Fix extra_port
The code turns SO_REUSEADDR on and off for the catchall socket
that grabs all arriving packets that aren't caught by a socket
for each individual IP Address.

The bug was that I wasn't keeping track of the new catchall
sockets for the extra port so it was turning SO_REUSEADDR on/off
for the socket for the wrong port.