ntpviz: Fix timezone handling.
The attempt to fix the deprecation warning for utcfromtimestamp
resulted in two different versions that didn't work in Python 2.  This
uses the last version with a fallback to the original version in
Python 2.

TESTED:
Tested a copy of this code fragment in Python 2.6-2.7 and 3.3-3.13.
Did not test the full ntpviz since it takes too long.