Adds attic tool for investigating packet timestamp behavior.
This tool tries all available timestamp options on the system, as well
as reporting basic info about the system and the struct timeXXX
formats.  It's a completely self-contained program that can be built
with a simple 'make' command.

Also, a simple (optional) Makefile is added to make it easy to build
programs with MacportsLegacySupport, mainly to get clock_gettime() on
OS versions earlier than 10.12.

Also adds a suitable .gitignore, while removing the minimal entry from
the parent .gitignore.

TESTED:
Produces plausible results on a wide variety of systems.