ntp_control.h: Fix excessive alignment requirement.
According to both the comment and the referenced issue, the
requirement is to be aligned to a 32-*bit* boundary, but the argument
to the __attribute__((aligned())) is in *bytes*.  The correct
alignment value should thus be 4, not 32.

This caused a warning on OpenBSD 5.6, where the compiler was unable to
guarantee 32-byte alignment of stack variables.

Also fixes the misleading comment.  The issue isn't specifically about
32-bit platforms.

NOTE: Removing the alignment altogether doesn't currently cause the
build to fail on a Beaglebone here (Debian 7, gcc 4.6.3), so there
isn't solid proof that this addresses issue #277, but the updated
alignment matches the stated intent of the original fix.

TESTED:
Built and ran tests on OSX 10.4-10.5 ppc, 10.4-10.6 i386, 10.5-12.x
x86_64, 11.x-15.x arm64, Debian 7 arm, Debian 10 arm64, Ubuntu 14
x86_64, CentOS 7 x86_64, Fedora 25 x86_64, Ubuntu 22 x86_64, FreeBSD
10.3 x86_64, OpenBSD 5.6 x86_64, and OpenBSD i386.