ntpq: Fix for Python 2.
As noted in the comment:

The BrokenPipeError doesn't exist in Python 2, but the problem that
requires catching it also doesn't seem to exist in Python 2, so just
make it a dummy in Python 2.

This just restores the old Python 2 behavior, without suppressing the
"Bailing out..." message.

TESTED:
Tested with Python 2.7 and 3.3-3.13.