Initialize `aes_key` variable as an array of bytes
Fedora 42 x86_64 ships with a gcc 15.x snapshot that produces the following warning:
```
../../tests/libntp/authkeys.c:12:35: error: initializer-string for array of ‘unsigned char’ truncates NUL terminator but destination lacks ‘nonstring’ attribute (17 chars into 16 available) [-Werror=unterminated-string-initialization]
   12 | const unsigned char aes_key[16] = "0123456789abcdef";
      |
```

We could use `__attribute__((nonstring))` to annotate the variable, but this is
only supported by gcc and not by clang.

I've tested this change on Fedora 42 x86_64 with gcc 15.0.1 and clang 20.1.2
and there were no errors/warnings, and all tests passed.

NTPsec/ntpsec#847